+++ ms2query/create_new_library/train_models.py
@@ -4,13 +4,15 @@ new models
 """
 
 import os
+from ms2deepscore import SettingsMS2Deepscore
+from ms2deepscore.train_new_model.train_ms2deepscore import train_ms2ds_model
 from spec2vec.model_building import train_new_word2vec_model
 from ms2query.clean_and_filter_spectra import (
     clean_normalize_and_split_annotated_spectra, create_spectrum_documents)
 from ms2query.create_new_library.library_files_creator import \
     LibraryFilesCreator
-from ms2query.create_new_library.train_ms2deepscore import \
-    train_ms2deepscore_wrapper
+from ms2query.create_new_library.split_data_for_training import \
+    split_spectra_on_inchikeys
 from ms2query.create_new_library.train_ms2query_model import (
     convert_to_onnx_model, train_ms2query_model)
 from ms2query.utils import load_matchms_spectrum_objects_from_file
@@ -20,17 +22,27 @@ class SettingsTrainingModels:
     def __init__(self,
                  settings: dict = None):
         default_settings = {"ms2ds_fraction_validation_spectra": 30,
-                            "ms2ds_epochs": 150,
                             "spec2vec_iterations": 30,
                             "ms2query_fraction_for_making_pairs": 40,
-                            "add_compound_classes": True}
+                            "add_compound_classes": True,
+                            "ms2ds_training_settings": SettingsMS2Deepscore(
+                                history_plot_file_name="ms2deepscore_training_history.svg",
+                                model_file_name="ms2deepscore_model.pt",
+                                epochs=150,
+                                embedding_dim=400,
+                                base_dims=(500, 500),
+                                min_mz=10,
+                                max_mz=1000,
+                                mz_bin_width=0.1,
+                                intensity_scaling=0.5
+                            )}
         if settings:
             for setting in settings:
                 assert setting in default_settings, \
                     f"Available settings are {default_settings.keys()}"
                 default_settings[setting] = settings[setting]
         self.ms2ds_fraction_validation_spectra: float = default_settings["ms2ds_fraction_validation_spectra"]
-        self.ms2ds_epochs: int = default_settings["ms2ds_epochs"]
+        self.ms2ds_training_settings: SettingsMS2Deepscore = default_settings["ms2ds_training_settings"]
         self.ms2query_fraction_for_making_pairs: int = default_settings["ms2query_fraction_for_making_pairs"]
         self.spec2vec_iterations = default_settings["spec2vec_iterations"]
         self.add_compound_classes: bool = default_settings["add_compound_classes"]
@@ -43,18 +55,15 @@ def train_all_models(annotated_training_spectra,
     if not os.path.isdir(output_folder):
         os.mkdir(output_folder)
     # set file names of new generated files
-    ms2deepscore_model_file_name = os.path.join(output_folder, "ms2deepscore_model.hdf5")
     spec2vec_model_file_name = os.path.join(output_folder, "spec2vec_model.model")
     ms2query_model_file_name = os.path.join(output_folder, "ms2query_model.onnx")
-    ms2ds_history_figure_file_name = os.path.join(output_folder, "ms2deepscore_training_history.svg")
 
     # Train MS2Deepscore model
-    train_ms2deepscore_wrapper(annotated_training_spectra,
-                               ms2deepscore_model_file_name,
-                               settings.ms2ds_fraction_validation_spectra,
-                               settings.ms2ds_epochs,
-                               ms2ds_history_figure_file_name
-                               )
+    training_spectra, validation_spectra = split_spectra_on_inchikeys(annotated_training_spectra,
+                                                                      settings.ms2ds_fraction_validation_spectra,
+                                                                      )
+    train_ms2ds_model(training_spectra, validation_spectra, output_folder,
+                      settings.ms2ds_training_settings)
 
     # Train Spec2Vec model
     spectrum_documents = create_spectrum_documents(annotated_training_spectra + unannotated_training_spectra)
@@ -68,7 +77,7 @@ def train_all_models(annotated_training_spectra,
     # Train MS2Query model
     ms2query_model = train_ms2query_model(annotated_training_spectra,
                                           os.path.join(output_folder, "library_for_training_ms2query"),
-                                          ms2deepscore_model_file_name,
+                                          os.path.join(output_folder, "ms2deepscore_model.pt"),
                                           spec2vec_model_file_name,
                                           fraction_for_training=settings.ms2query_fraction_for_making_pairs)
     convert_to_onnx_model(ms2query_model, ms2query_model_file_name)
@@ -77,7 +86,7 @@ def train_all_models(annotated_training_spectra,
     library_files_creator = LibraryFilesCreator(annotated_training_spectra,
                                                 output_folder,
                                                 spec2vec_model_file_name,
-                                                ms2deepscore_model_file_name,
+                                                os.path.join(output_folder, "ms2deepscore_model.pt"),
                                                 add_compound_classes=settings.add_compound_classes)
     library_files_creator.create_all_library_files()
 
@@ -85,7 +94,7 @@ def train_all_models(annotated_training_spectra,
 def clean_and_train_models(spectrum_file: str,
                            ion_mode: str,
                            output_folder,
-                           model_train_settings = None,
+                           model_train_settings=None,
                            do_pubchem_lookup = True):
     """Trains a new MS2Deepscore, Spec2Vec and MS2Query model and creates all needed library files